pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "." }
packages = ["mbu_dev_shared_components"]

[project]
name = "mbu_dev_shared_components"
version = "2.4.1"  # Specify the version manually here
authors = [
  { name="MBU", email="rpa@mbu.aarhus.dk" },
]
description = "Shared components to use in RPA projects"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
]
license = "MIT"
dependencies = [
  "openpyxl >= 3.1.2",
  "pyodbc >= 5.1.0",
  "requests_ntlm >= 1.2.0",
  "python-dateutil == 2.9.*",
  "cryptography >= 43.0.0",
  "office365-rest-python-client",
  "uiautomation",
  "pillow",
  "psutil",
  "docx2pdf",
  "rawpy",
]

[project.optional-dependencies]
dev = [
  "pytest >= 7.0",
]
